FIxes and improvements
